Home: move the shared solution data layer into home/solutions/
The datasource probes, PromQL helpers and telemetry queries are about solutions,
not about the recommendations that happened to be their first consumer. Moving
them out of Recommendations/ lets the solution definitions that follow import
them without reaching sideways into another section's directory.

Pure move: file contents are unchanged and only import paths are rewritten.
