+++ public/app/features/home/solutions/SolutionSparkline.tsx
