+++ public/app/features/home/solutions/appPluginIds.ts
