+++ public/app/features/home/solutions/kubernetesData.ts
