+++ public/app/features/home/solutions/probeUtils.ts
