+++ public/app/features/home/solutions/promQuery.ts
