+++ public/app/features/home/solutions/solutionState.test.ts
