+++ public/app/features/home/solutions/telemetryData.test.ts
