+++ public/app/features/home/solutions/telemetryData.ts
